fix: make Simple Home progress value explicitly double

+++ lib/screens/simple_home_screen.dart
@@ -158,7 +158,7 @@ class _SimpleHomeScreenState extends State<SimpleHomeScreen> {
               surahName: lastRead['surahName']?.toString() ?? '',
               paraNo: lastRead['paraNo'] is int ? lastRead['paraNo'] as int : int.tryParse('${lastRead['paraNo']}') ?? 1,
               pageNo: lastRead['pageNo'] is int ? lastRead['pageNo'] as int : int.tryParse('${lastRead['pageNo']}') ?? 1,
-              progress: (lastRead['progress'] is num ? (lastRead['progress'] as num).toDouble() : 0).clamp(0.0, 1.0),
+              progress: ((lastRead['progress'] is num ? (lastRead['progress'] as num).toDouble() : 0).clamp(0.0, 1.0)).toDouble(),
               languageCode: languageCode,
               onTap: () => _open(const OnudhabonQuranScreen(openLastRead: true)),
             ),
